# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Payment Gateway API
  version: 2.0.0
  description: |
    Secure Payment Gateway API for Merchants. Supports payment processing,
    wallet management, webhook notifications, and merchant self-service.
    
    **Authentication Methods:**
    - HMAC-SHA256: For payment operations (machine-to-machine)
    - JWT Bearer: For dashboard and management (human interaction)
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.gateway.com/api/v1
    description: Production
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Merchant-Access-Key
    SignatureAuth:
      type: apiKey
      in: header
      name: X-Signature # HMAC-SHA256 signature
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransactionResponse:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        reference_id:
          type: string
        status:
          type: string
          enum: [SUCCESS, FAILED, PENDING]
        amount:
          type: integer
        currency:
          type: string
        transaction_type:
          type: string
          enum: [PAYMENT, REFUND, TOPUP]
        processed_at:
          type: string
          format: date-time

    # --- Auth ---
...
paths:
...
  /payments/{id}/status:
    get:
      tags: [Payments]
      summary: Get payment transaction status
      description: Returns the current status of a specific transaction.
      operationId: getPaymentStatus
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Transaction UUID
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "404":
          description: Transaction not found

  # ----------------------------------------------------------
  # MERCHANT MANAGEMENT (JWT auth)
  # ----------------------------------------------------------
...